+++ transaction/chaintracker/chaintracker.go
@@ -1,7 +1,12 @@
 package chaintracker
 
-import "github.com/bsv-blockchain/go-sdk/chainhash"
+import (
+	"context"
+
+	"github.com/bsv-blockchain/go-sdk/chainhash"
+)
 
 type ChainTracker interface {
-	IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool, error)
+	IsValidRootForHeight(ctx context.Context, root *chainhash.Hash, height uint32) (bool, error)
+	CurrentHeight(ctx context.Context) (uint32, error)
 }