+++ transaction/chaintracker/headers_client/headers_client.go
@@ -34,7 +34,7 @@ type Client struct {
 	ApiKey string
 }
 
-func (c Client) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool, error) {
+func (c Client) IsValidRootForHeight(ctx context.Context, root *chainhash.Hash, height uint32) (bool, error) {
 	type requestBody struct {
 		MerkleRoot  string `json:"merkleRoot"`
 		BlockHeight uint32 `json:"blockHeight"`
@@ -46,7 +46,7 @@ func (c Client) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool,
 		return false, fmt.Errorf("error marshaling JSON: %v", err)
 	}
 
-	req, err := http.NewRequest("POST", c.Url+"/api/v1/chain/merkleroot/verify", bytes.NewBuffer(jsonPayload))
+	req, err := http.NewRequestWithContext(ctx, "POST", c.Url+"/api/v1/chain/merkleroot/verify", bytes.NewBuffer(jsonPayload))
 	if err != nil {
 		return false, fmt.Errorf("error creating request: %v", err)
 	}
@@ -59,7 +59,7 @@ func (c Client) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool,
 	if err != nil {
 		return false, fmt.Errorf("error sending request: %v", err)
 	}
-	defer func() { _ = resp.Body.Close() }()
+	defer resp.Body.Close()
 
 	body, err := io.ReadAll(resp.Body)
 	if err != nil {
@@ -90,7 +90,7 @@ func (c *Client) BlockByHeight(ctx context.Context, height uint32) (*Header, err
 	if res, err := client.Do(req); err != nil {
 		return nil, err
 	} else {
-		defer func() { _ = res.Body.Close() }()
+		defer res.Body.Close()
 		if err := json.NewDecoder(res.Body).Decode(&headers); err != nil {
 			return nil, err
 		}
@@ -119,7 +119,7 @@ func (c *Client) GetBlockState(ctx context.Context, hash string) (*State, error)
 	if res, err := client.Do(req); err != nil {
 		return nil, err
 	} else {
-		defer func() { _ = res.Body.Close() }()
+		defer res.Body.Close()
 		if err := json.NewDecoder(res.Body).Decode(headerState); err != nil {
 			return nil, err
 		}
@@ -138,10 +138,18 @@ func (c *Client) GetChaintip(ctx context.Context) (*State, error) {
 	if res, err := client.Do(req); err != nil {
 		return nil, err
 	} else {
-		defer func() { _ = res.Body.Close() }()
+		defer res.Body.Close()
 		if err := json.NewDecoder(res.Body).Decode(headerState); err != nil {
 			return nil, err
 		}
 	}
 	return headerState, nil
+}
+
+func (c *Client) CurrentHeight(ctx context.Context) (uint32, error) {
+	tip, err := c.GetChaintip(ctx)
+	if err != nil {
+		return 0, err
+	}
+	return tip.Height, nil
 }