+++ transaction/chaintracker/whatsonchain.go
@@ -34,6 +34,10 @@ type WhatsOnChain struct {
 	client  *http.Client
 }
 
+type ChainInfo struct {
+	Blocks uint32 `json:"blocks"`
+}
+
 func NewWhatsOnChain(network Network, apiKey string) *WhatsOnChain {
 	return &WhatsOnChain{
 		Network: network,
@@ -44,9 +48,9 @@ func NewWhatsOnChain(network Network, apiKey string) *WhatsOnChain {
 }
 
 // Assuming BlockHeader is defined elsewhere
-func (w *WhatsOnChain) GetBlockHeader(height uint32) (header *BlockHeader, err error) {
+func (w *WhatsOnChain) GetBlockHeader(ctx context.Context, height uint32) (header *BlockHeader, err error) {
 	url := fmt.Sprintf("%s/block/%d/header", w.baseURL, height)
-	req, err := http.NewRequestWithContext(context.Background(), "GET", url, nil)
+	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
 	if err != nil {
 		return nil, err
 	}
@@ -57,7 +61,7 @@ func (w *WhatsOnChain) GetBlockHeader(height uint32) (header *BlockHeader, err e
 	if err != nil {
 		return nil, err
 	}
-	defer func() { _ = resp.Body.Close() }()
+	defer resp.Body.Close()
 
 	if resp.StatusCode == http.StatusNotFound {
 		return nil, nil
@@ -74,10 +78,41 @@ func (w *WhatsOnChain) GetBlockHeader(height uint32) (header *BlockHeader, err e
 	return header, nil
 }
 
-func (w *WhatsOnChain) IsValidRootForHeight(root *chainhash.Hash, height uint32) (bool, error) {
-	if header, err := w.GetBlockHeader(height); err != nil {
+func (w *WhatsOnChain) IsValidRootForHeight(ctx context.Context, root *chainhash.Hash, height uint32) (bool, error) {
+	if header, err := w.GetBlockHeader(ctx, height); err != nil {
 		return false, err
 	} else {
 		return header.MerkleRoot.IsEqual(root), nil
 	}
+}
+
+// Assuming BlockHeader is defined elsewhere
+func (w *WhatsOnChain) CurrentHeight(ctx context.Context) (height uint32, err error) {
+	url := fmt.Sprintf("%s/chain/info", w.baseURL)
+	req, err := http.NewRequestWithContext(ctx, "GET", url, nil)
+	if err != nil {
+		return
+	}
+
+	req.Header.Set("Authorization", w.ApiKey)
+
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return
+	}
+	defer resp.Body.Close()
+
+	if resp.StatusCode == http.StatusNotFound {
+		return 0, fmt.Errorf("chain info not found for network %s", w.Network)
+	}
+	if resp.StatusCode != http.StatusOK {
+		return 0, fmt.Errorf("failed to verify merkleroot for height %d: %v", height, resp.Status)
+	}
+
+	info := &ChainInfo{}
+	if err := json.NewDecoder(resp.Body).Decode(info); err != nil {
+		return 0, err
+	}
+
+	return info.Blocks, nil
 }